app/requirements.txt
fastapi
pillow
tesseract
pytesseract
pdf2image
